First pass at extracting refund summary logic to its own ViewModel
