+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/payments/refunds/IssueRefundViewModel.kt
@@ -5,25 +5,14 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.asLiveData
 import androidx.lifecycle.viewModelScope
 import com.woocommerce.android.R
-import com.woocommerce.android.WooException
 import com.woocommerce.android.analytics.AnalyticsEvent.CREATE_ORDER_REFUND_ITEM_QUANTITY_DIALOG_OPENED
 import com.woocommerce.android.analytics.AnalyticsEvent.CREATE_ORDER_REFUND_NEXT_BUTTON_TAPPED
 import com.woocommerce.android.analytics.AnalyticsEvent.CREATE_ORDER_REFUND_SELECT_ALL_ITEMS_BUTTON_TAPPED
-import com.woocommerce.android.analytics.AnalyticsEvent.CREATE_ORDER_REFUND_SUMMARY_REFUND_BUTTON_TAPPED
-import com.woocommerce.android.analytics.AnalyticsEvent.ORDER_NOTE_ADD_FAILED
-import com.woocommerce.android.analytics.AnalyticsEvent.ORDER_NOTE_ADD_SUCCESS
-import com.woocommerce.android.analytics.AnalyticsEvent.REFUND_CREATE
-import com.woocommerce.android.analytics.AnalyticsEvent.REFUND_CREATE_FAILED
-import com.woocommerce.android.analytics.AnalyticsEvent.REFUND_CREATE_SUCCESS
 import com.woocommerce.android.analytics.AnalyticsTracker
 import com.woocommerce.android.analytics.AnalyticsTrackerWrapper
 import com.woocommerce.android.extensions.adminUrlOrDefault
-import com.woocommerce.android.extensions.isCashPayment
-import com.woocommerce.android.extensions.isEqualTo
 import com.woocommerce.android.model.Order
 import com.woocommerce.android.model.OrderMapper
-import com.woocommerce.android.model.OrderNote
-import com.woocommerce.android.model.PaymentGateway
 import com.woocommerce.android.model.Refund
 import com.woocommerce.android.model.getMaxRefundQuantities
 import com.woocommerce.android.model.toAppModel
@@ -32,23 +21,18 @@ import com.woocommerce.android.tools.SelectedSite
 import com.woocommerce.android.ui.orders.details.OrderDetailRepository
 import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.IssueRefundEvent.OpenUrl
 import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.IssueRefundEvent.ShowNumberPicker
-import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.IssueRefundEvent.ShowRefundConfirmation
-import com.woocommerce.android.ui.payments.refunds.IssueRefundViewModel.IssueRefundEvent.ShowRefundSummary
 import com.woocommerce.android.util.CoroutineDispatchers
 import com.woocommerce.android.util.CurrencyFormatter
 import com.woocommerce.android.util.max
 import com.woocommerce.android.util.min
 import com.woocommerce.android.viewmodel.LiveDataDelegate
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event
-import com.woocommerce.android.viewmodel.MultiLiveEvent.Event.Exit
-import com.woocommerce.android.viewmodel.MultiLiveEvent.Event.ShowSnackbar
 import com.woocommerce.android.viewmodel.ResourceProvider
 import com.woocommerce.android.viewmodel.ScopedViewModel
 import com.woocommerce.android.viewmodel.getNullableListStateFlow
 import com.woocommerce.android.viewmodel.getStateFlow
 import com.woocommerce.android.viewmodel.navArgs
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.SharedFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.combine
@@ -60,11 +44,7 @@ import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.flow.shareIn
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import kotlinx.coroutines.withContext
 import kotlinx.parcelize.Parcelize
-import org.wordpress.android.fluxc.model.refunds.RefundRequestItem
-import org.wordpress.android.fluxc.model.refunds.WCRefundModel
-import org.wordpress.android.fluxc.network.rest.wpcom.wc.WooResult
 import org.wordpress.android.fluxc.store.WCGatewayStore
 import org.wordpress.android.fluxc.store.WCOrderStore
 import org.wordpress.android.fluxc.store.WCRefundStore
@@ -90,10 +70,6 @@ class IssueRefundViewModel @Inject constructor(
     private val analyticsTrackerWrapper: AnalyticsTrackerWrapper,
     private val coroutineDispatchers: CoroutineDispatchers
 ) : ScopedViewModel(savedState) {
-    companion object {
-        private const val REFUND_METHOD_MANUAL = "manual"
-    }
-
     private val orderFlow: SharedFlow<Order> = flow {
         val order = requireNotNull(
             orderStore.getOrderByIdAndSite(arguments.orderId, selectedSite.get())?.let { orderMapper.toAppModel(it) }
@@ -185,16 +161,7 @@ class IssueRefundViewModel @Inject constructor(
     @Suppress("OPT_IN_USAGE")
     val commonStateLiveData = LiveDataDelegate(savedState, CommonViewState())
 
-    /**
-     * Saving more data than necessary into the SavedState has associated risks which were not known at the time this
-     * field was implemented - after we ensure we don't save unnecessary data, we can replace @Suppress("OPT_IN_USAGE")
-     * with @OptIn(LiveDelegateSavedStateAPI::class).
-     */
-    @Suppress("OPT_IN_USAGE")
-    val refundSummaryStateLiveData = LiveDataDelegate(savedState, RefundSummaryViewState())
-
     private var commonState by commonStateLiveData
-    private var refundSummaryState by refundSummaryStateLiveData
 
     private val order: Order
         get() = requireNotNull(orderFlow.replayCache.firstOrNull()) {
@@ -213,29 +180,19 @@ class IssueRefundViewModel @Inject constructor(
         get() = allShippingLineIds.filterNot { shippingId ->
             refunds.any { it.id == shippingId }
         }
-    private val Order.maxRefund: BigDecimal
-        get() = total - refundTotal
 
     private val Order.containsOnlyCustomAmounts: Boolean
         get() = items.isEmpty() && shippingLines.isEmpty() && feesLines.isNotEmpty()
 
     private val formatCurrency: (BigDecimal) -> String
         get() = currencyFormatter.buildBigDecimalFormatter(order.currency)
-    private lateinit var gateway: PaymentGateway
-    private var cardType = PaymentMethodType.CARD_PRESENT
     private val arguments: RefundsArgs by savedState.navArgs()
-
-    private var refundJob: Job? = null
-    val isRefundInProgress: Boolean
-        get() = refundJob?.isActive ?: false
 
     init {
         refunds = refundStore.getAllRefunds(selectedSite.get(), arguments.orderId).map { it.toAppModel() }
 
         viewModelScope.launch {
-            gateway = loadPaymentGateway()
             initRefundItems()
-            initRefundSummaryState()
         }
     }
 
@@ -361,31 +318,6 @@ class IssueRefundViewModel @Inject constructor(
         )
     }
 
-    private fun initRefundSummaryState() {
-        if (refundSummaryStateLiveData.hasInitialValue) {
-            val manualRefundMethod = resourceProvider.getString(R.string.order_refunds_manual_refund)
-            if (!order.paymentMethod.isCashPayment && (!gateway.isEnabled || !gateway.supportsRefunds)) {
-                val paymentTitle = if (gateway.title.isNotBlank()) {
-                    resourceProvider.getString(R.string.order_refunds_method, manualRefundMethod, gateway.title)
-                } else {
-                    manualRefundMethod
-                }
-                updateRefundSummaryState(paymentTitle, isMethodDescriptionVisible = true)
-            } else {
-                enrichRefundMethodWithCardDetails(gateway.title.ifBlank { manualRefundMethod })
-            }
-        }
-    }
-
-    private suspend fun loadPaymentGateway(): PaymentGateway = withContext(coroutineDispatchers.io) {
-        val paymentGateway = gatewayStore.getGateway(selectedSite.get(), orderFlow.first().paymentMethod)?.toAppModel()
-        return@withContext if (paymentGateway != null && paymentGateway.isEnabled) {
-            paymentGateway
-        } else {
-            PaymentGateway(methodTitle = REFUND_METHOD_MANUAL)
-        }
-    }
-
     fun onNextButtonTappedFromItems() {
         analyticsTrackerWrapper.track(
             CREATE_ORDER_REFUND_NEXT_BUTTON_TAPPED,
@@ -402,193 +334,12 @@ class IssueRefundViewModel @Inject constructor(
     }
 
     private fun showRefundSummary() {
-        refundSummaryState = refundSummaryState.copy(
-            isFormEnabled = true,
-            previouslyRefunded = formatCurrency(order.refundTotal),
-            refundAmount = commonState.refundTotal,
-            refundAmountFormatted = formatCurrency(commonState.refundTotal)
-        )
-
-        triggerEvent(ShowRefundSummary)
-    }
-
-    fun onRefundConfirmed(wasConfirmed: Boolean) {
-        if (wasConfirmed) {
-            if (networkStatus.isConnected()) {
-                refundJob = launch {
-                    refundSummaryState = refundSummaryState.copy(
-                        isFormEnabled = false
-                    )
-                    if (isInteracRefund()) {
-                        triggerEvent(IssueRefundEvent.NavigateToCardReaderScreen(order.id, commonState.refundTotal))
-                    } else {
-                        triggerEvent(
-                            ShowSnackbar(
-                                R.string.order_refunds_amount_refund_progress_message,
-                                arrayOf(formatCurrency(commonState.refundTotal))
-                            )
-                        )
-                        refund()
-                    }
-
-                    analyticsTrackerWrapper.track(
-                        REFUND_CREATE,
-                        mapOf(
-                            AnalyticsTracker.KEY_ORDER_ID to order.id,
-                            AnalyticsTracker.KEY_REFUND_IS_FULL to
-                                (commonState.refundTotal isEqualTo order.maxRefund).toString(),
-                            AnalyticsTracker.KEY_REFUND_METHOD to gateway.methodTitle,
-                            AnalyticsTracker.KEY_AMOUNT to commonState.refundTotal.toString()
-                        )
-                    )
-                    refundSummaryState = refundSummaryState.copy(isFormEnabled = true)
-                }
-            } else {
-                triggerEvent(ShowSnackbar(R.string.offline_error))
-            }
-        }
-    }
-
-    private fun isInteracRefund() = cardType == PaymentMethodType.INTERAC_PRESENT
-
-    /*
-       This method does the actual refund in case of non-interac refund. In case of Interac refund, the actual
-       refund happens on the client-side and this method updates the WCPay backend about the refund success status and
-       does not process the refund itself.
-
-       For non-Interac refund -> Process the refund (Entire refund logic lives in the backend)
-       For Interac refund -> Update the backend of the successful refund. The actual refund happens on the client-side
-     */
-    fun refund() {
-        triggerUIMessageIfRefundIsInterac()
-        launch {
-            val result = initiateRefund()
-            if (result.isError) {
-                trackRefundError(result)
-                triggerUIMessage()
-            } else {
-                trackRefundSuccess(result)
-                updateRefundSummaryStateWithOrderNote()
-                triggerEvent(ShowSnackbar(R.string.order_refunds_amount_refund_successful))
-                triggerEvent(Exit)
-            }
-        }
-    }
-
-    private fun triggerUIMessageIfRefundIsInterac() {
-        if (isInteracRefund()) {
-            triggerEvent(ShowSnackbar(R.string.card_reader_interac_refund_notifying_backend_about_successful_refund))
-        }
-    }
-
-    private suspend fun initiateRefund(): WooResult<WCRefundModel> {
-        val allItems = mutableListOf<RefundRequestItem>()
-        refundItems.value?.let {
-            it.forEach { item -> allItems.add(item.toDataModel()) }
-        }
-
-        val selectedShipping = if (isShippingMainSwitchChecked.value) {
-            refundShippingLines.value?.filter { it.isSelected }
-        } else {
-            emptyList()
-        }
-        selectedShipping?.forEach { allItems.add(it.toDataModel()) }
-
-        val selectedFees = if (isFeesMainSwitchChecked.value) {
-            refundFeeLines.value?.filter { it.isSelected }
-        } else {
-            emptyList()
-        }
-        selectedFees?.forEach { allItems.add(it.toDataModel()) }
-
-        return refundStore.createItemsRefund(
-            site = selectedSite.get(),
-            orderId = order.id,
-            amount = refundSummaryState.refundAmount,
-            reason = refundSummaryState.refundReason ?: "",
-            restockItems = true,
-            autoRefund = gateway.supportsRefunds,
-            items = allItems
-        )
-    }
-
-    private fun trackRefundError(result: WooResult<WCRefundModel>) {
-        analyticsTrackerWrapper.track(
-            REFUND_CREATE_FAILED,
-            mapOf(
-                AnalyticsTracker.KEY_ORDER_ID to order.id,
-                AnalyticsTracker.KEY_ERROR_CONTEXT to this::class.java.simpleName,
-                AnalyticsTracker.KEY_ERROR_TYPE to result.error.type.toString(),
-                AnalyticsTracker.KEY_ERROR_DESC to result.error.message
-            )
-        )
-    }
-
-    private fun triggerUIMessage() {
-        if (isInteracRefund()) {
-            triggerEvent(
-                ShowSnackbar(
-                    R.string.card_reader_interac_refund_notifying_backend_about_successful_refund_failed
-                )
-            )
-        } else {
-            triggerEvent(ShowSnackbar(R.string.order_refunds_amount_refund_error))
-        }
-    }
-
-    private fun trackRefundSuccess(result: WooResult<WCRefundModel>) {
-        analyticsTrackerWrapper.track(
-            REFUND_CREATE_SUCCESS,
-            mapOf(
-                AnalyticsTracker.KEY_ORDER_ID to order.id,
-                AnalyticsTracker.KEY_ID to result.model?.id
-            )
-        )
-    }
-
-    private suspend fun updateRefundSummaryStateWithOrderNote() {
-        refundSummaryState.refundReason?.let { reason ->
-            if (reason.isNotBlank()) {
-                addOrderNote(reason)
-            }
-        }
-    }
-
-    private suspend fun addOrderNote(reason: String) {
-        val note = OrderNote(note = reason, isCustomerNote = false)
-        orderDetailRepository.addOrderNote(order.id, note).fold(
-            onSuccess = {
-                analyticsTrackerWrapper.track(ORDER_NOTE_ADD_SUCCESS)
-            },
-            onFailure = {
-                analyticsTrackerWrapper.track(
-                    ORDER_NOTE_ADD_FAILED,
-                    prepareTracksEventsDetails(it as WooException)
-                )
-            }
-        )
-    }
-
-    fun onRefundIssued(reason: String) = viewModelScope.launch {
-        analyticsTrackerWrapper.track(
-            CREATE_ORDER_REFUND_SUMMARY_REFUND_BUTTON_TAPPED,
-            mapOf(
-                AnalyticsTracker.KEY_ORDER_ID to order.id
-            )
-        )
-
-        refundSummaryState = refundSummaryState.copy(
-            refundReason = reason
-        )
-
         triggerEvent(
-            ShowRefundConfirmation(
-                resourceProvider.getString(
-                    R.string.order_refunds_title_with_amount,
-                    formatCurrency(commonState.refundTotal)
-                ),
-                resourceProvider.getString(R.string.order_refunds_confirmation),
-                resourceProvider.getString(R.string.order_refunds_refund)
+            IssueRefundEvent.ShowRefundSummary(
+                orderId = arguments.orderId,
+                refundItems = refundItems.value.orEmpty() +
+                    refundShippingLines.value.orEmpty() +
+                    refundFeeLines.value.orEmpty()
             )
         )
     }
@@ -602,14 +353,6 @@ class IssueRefundViewModel @Inject constructor(
             CREATE_ORDER_REFUND_ITEM_QUANTITY_DIALOG_OPENED,
             mapOf(AnalyticsTracker.KEY_ORDER_ID to order.id)
         )
-    }
-
-    /**
-     * Checks if the refund summary button label should be enabled. If the max length for the text field is
-     * surpassed, the button should be disabled until the text is brought within the maximum length.
-     */
-    fun onRefundSummaryTextChanged(maxLength: Int, currLength: Int) {
-        refundSummaryState = refundSummaryState.copy(isSummaryTextTooLong = currLength > maxLength)
     }
 
     fun onRefundQuantityChanged(uniqueId: Long, newQuantity: Int) {
@@ -690,52 +433,6 @@ class IssueRefundViewModel @Inject constructor(
         }
     }
 
-    private fun enrichRefundMethodWithCardDetails(refundMethod: String) {
-        val chargeId = order.chargeId
-        if (chargeId != null) {
-            loadCardDetails(chargeId, refundMethod)
-        } else {
-            updateRefundSummaryState(refundMethod, isMethodDescriptionVisible = false)
-        }
-    }
-
-    private fun loadCardDetails(chargeId: String, refundMethod: String) {
-        launch {
-            refundSummaryState = refundSummaryState.copy(isFetchingCardData = true)
-            val result = paymentChargeRepository.fetchCardDataUsedForOrderPayment(chargeId)
-            refundSummaryState = refundSummaryState.copy(isFetchingCardData = false)
-            when (result) {
-                is PaymentChargeRepository.CardDataUsedForOrderPaymentResult.Success -> {
-                    cardType = PaymentMethodType.fromValue(result.paymentMethodType ?: "card_present")
-                    val refundMethodWithCard = result.run {
-                        val brand = result.cardBrand.orEmpty().replaceFirstChar { it.uppercase() }
-                        val last4 = result.cardLast4.orEmpty()
-                        "$refundMethod ($brand **** $last4)"
-                    }
-                    updateRefundSummaryState(refundMethodWithCard, isMethodDescriptionVisible = false)
-                }
-
-                PaymentChargeRepository.CardDataUsedForOrderPaymentResult.Error -> {
-                    cardType = PaymentMethodType.CARD_PRESENT
-                    updateRefundSummaryState(refundMethod, isMethodDescriptionVisible = false)
-                }
-            }
-        }
-    }
-
-    private fun updateRefundSummaryState(refundMethod: String, isMethodDescriptionVisible: Boolean) {
-        refundSummaryState = refundSummaryState.copy(
-            refundMethod = refundMethod,
-            isMethodDescriptionVisible = isMethodDescriptionVisible
-        )
-    }
-
-    private fun prepareTracksEventsDetails(exception: WooException) = mapOf(
-        AnalyticsTracker.KEY_ERROR_CONTEXT to this::class.java.simpleName,
-        AnalyticsTracker.KEY_ERROR_TYPE to exception.error.type.toString(),
-        AnalyticsTracker.KEY_ERROR_DESC to exception.error.message
-    )
-
     data class RefundByItemsViewState(
         val currency: String,
         val productsSection: ProductsRefundSection,
@@ -792,22 +489,6 @@ class IssueRefundViewModel @Inject constructor(
     )
 
     @Parcelize
-    data class RefundSummaryViewState(
-        val isFormEnabled: Boolean? = null,
-        val previouslyRefunded: String? = null,
-        val refundAmount: BigDecimal? = null,
-        val refundAmountFormatted: String? = null,
-        val refundMethod: String? = null,
-        val refundReason: String? = null,
-        val isMethodDescriptionVisible: Boolean? = null,
-        val isSummaryTextTooLong: Boolean = false,
-        val isFetchingCardData: Boolean = false,
-    ) : Parcelable {
-        val isSubmitButtonEnabled: Boolean
-            get() = !isSummaryTextTooLong && !isFetchingCardData
-    }
-
-    @Parcelize
     data class CommonViewState(
         val refundTotal: BigDecimal = BigDecimal.ZERO,
         val screenTitle: String? = null
@@ -821,9 +502,11 @@ class IssueRefundViewModel @Inject constructor(
             val confirmButtonTitle: String
         ) : IssueRefundEvent()
 
-        data object ShowRefundSummary : IssueRefundEvent()
         data class OpenUrl(val url: String) : IssueRefundEvent()
-        data class NavigateToCardReaderScreen(val orderId: Long, val refundAmount: BigDecimal) : IssueRefundEvent()
+        data class ShowRefundSummary(
+            val orderId: Long,
+            val refundItems: List<RefundItem>
+        ) : IssueRefundEvent()
     }
 
     enum class PaymentMethodType(val paymentMethodType: String) {
